requirements.txt
streamlit~=1.41.1
tensorflow~=2.18.0
numpy~=2.0.2
pandas~=2.2.3
pillow~=11.1.0
opencv-python~=4.11.0.86